+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Breaking changes to `unstable-v3`
+
+* Change `Shape` trait methods to return `impl Iterator`
 
 ## [2.0.1] - 2023-11-01